[FEATURE] Allow session transfer for file import
When the backend session cookie same site attribute is not set
to "none" browsers do not send the cookie when the request
is initiated from a different domain. To circumvent this,
a new session id is generated for easydb import only,
which is sent to the easydb server and then back to TYPO3
to authenticate easydb import requests.

This needs to be activated in extension settings, is deactivated
by default and will also be deactivated when not required (cookie
same site set to "none" or easdb host is same site).